+++ src/dbutil/Datastore.h
@@ -10,8 +10,8 @@ class BLTAbstractDataStore
 {
 public:
 	virtual ~BLTAbstractDataStore() {}
-	virtual size_t write(size_t position_in_file, uint8_t const *data, size_t length); // Stubbed with an abort
-	virtual size_t read(size_t position_in_file, uint8_t * data, size_t length) = 0;
+	virtual size_t write(uint64_t position_in_file, uint8_t const *data, size_t length); // Stubbed with an abort
+	virtual size_t read(uint64_t position_in_file, uint8_t * data, size_t length) = 0;
 	virtual bool close() = 0;
 	virtual size_t size() const = 0;
 	virtual bool is_asynchronous() const = 0;
@@ -25,7 +25,7 @@ class BLTFileDataStore : public BLTAbstractDataStore
 public:
 	BLTFileDataStore(std::string filePath);
 	virtual ~BLTFileDataStore();
-	virtual size_t read(size_t position_in_file, uint8_t * data, size_t length) override;
+	virtual size_t read(uint64_t position_in_file, uint8_t * data, size_t length) override;
 	virtual bool close() override;
 	virtual size_t size() const override;
 	virtual bool is_asynchronous() const override;
